Projeto finalizado. Features no README

+++ src/main/java/one/digitalinnovation/gof/service/impl/ClienteServiceImpl.java
@@ -6,18 +6,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import one.digitalinnovation.gof.model.Cliente;
-import one.digitalinnovation.gof.model.ClienteRepository;
+import one.digitalinnovation.gof.repository.ClienteRepository;
 import one.digitalinnovation.gof.model.Endereco;
-import one.digitalinnovation.gof.model.EnderecoRepository;
+import one.digitalinnovation.gof.repository.EnderecoRepository;
 import one.digitalinnovation.gof.service.ClienteService;
 import one.digitalinnovation.gof.service.ViaCepService;
+
 
 /**
  * Implementação da <b>Strategy</b> {@link ClienteService}, a qual pode ser
  * injetada pelo Spring (via {@link Autowired}). Com isso, como essa classe é um
  * {@link Service}, ela será tratada como um <b>Singleton</b>.
- * 
- * @author falvojr
+ *
  */
 @Service
 public class ClienteServiceImpl implements ClienteService {
@@ -43,12 +43,30 @@ public class ClienteServiceImpl implements ClienteService {
 	public Cliente buscarPorId(Long id) {
 		// Buscar Cliente por ID.
 		Optional<Cliente> cliente = clienteRepository.findById(id);
-		return cliente.get();
+		if(cliente.isPresent())
+		{
+			return cliente.get();
+		}
+		else
+		{
+			throw new IllegalArgumentException("There is no client!");
+		}
 	}
 
 	@Override
 	public void inserir(Cliente cliente) {
-		salvarClienteComCep(cliente);
+		//Optional<Cliente> clienteToSave = clienteRepository.findByName(cliente.getNome());
+		Optional<Cliente> clienteToSave = clienteRepository.findByNameAndCpf(cliente.getNome(), cliente.getCpf());
+		System.out.println("-----------------------------------------------------------> " + clienteToSave);
+		if(validaCpf(cliente)) {
+			if(clienteToSave.isEmpty()){
+				salvarClienteComCep(cliente);
+			} else {
+				throw new IllegalArgumentException("This client already exist");
+			}
+		} else {
+			throw new IllegalArgumentException("This cpf is invalid");
+		}
 	}
 
 	@Override
@@ -56,14 +74,25 @@ public class ClienteServiceImpl implements ClienteService {
 		// Buscar Cliente por ID, caso exista:
 		Optional<Cliente> clienteBd = clienteRepository.findById(id);
 		if (clienteBd.isPresent()) {
-			salvarClienteComCep(cliente);
+			if(validaCpf(cliente)){
+				atualizarClienteExistente(cliente);
+			} else {
+				throw new IllegalArgumentException("This cpf is invalid");
+			}
+		} else {
+			throw new IllegalArgumentException("This client does not exist");
 		}
 	}
 
 	@Override
 	public void deletar(Long id) {
 		// Deletar Cliente por ID.
-		clienteRepository.deleteById(id);
+		Optional<Cliente> cliente = clienteRepository.findById(id);
+		if(cliente.isPresent()) {
+			clienteRepository.deleteById(id);
+		} else {
+			throw new IllegalArgumentException("This client does not exist");
+		}
 	}
 
 	private void salvarClienteComCep(Cliente cliente) {
@@ -79,5 +108,65 @@ public class ClienteServiceImpl implements ClienteService {
 		// Inserir Cliente, vinculando o Endereco (novo ou existente).
 		clienteRepository.save(cliente);
 	}
+	private void atualizarClienteExistente(Cliente cliente) {
+		// Verificar se o Cliente já existe pelo ID.
+		Long clienteId = cliente.getId();
+		if (clienteId == null || !clienteRepository.existsById(clienteId)) {
+			throw new IllegalArgumentException("Client not found for update.");
+		}
+		// Atualizar o Cliente
+		salvarClienteComCep(cliente);
+	}
 
+	private boolean validaCpf(Cliente cliente){
+		String cpf = cliente.getCpf();
+		boolean response = false;
+		cpf = cpf.replaceAll("[^0-9]", "");
+
+		try {
+			int invalidcpf = 0;
+			int digit1 = 0;
+			int digit2 = 0;
+
+			while (cpf.length() < 11) {
+				cpf = "0" + cpf;
+			}
+
+			if (cpf.length() == 11) {
+				for (int i = 1; i <= cpf.length(); i++) {
+					if (cpf.charAt(0) == cpf.charAt(i - 1)) {
+						invalidcpf++;
+					}
+					if (i < 10) {
+						digit1 += Integer.parseInt(String.valueOf(cpf.charAt(i - 1))) * i;
+					}
+					if (i < 11) {
+						digit2 += Integer.parseInt(String.valueOf(cpf.charAt(i - 1))) * (i - 1);
+					}
+				}
+
+				if (invalidcpf == 11) {
+					System.out.println("CPF inválido! " + cpf);
+					return response;
+				}
+
+				digit1 = digit1 % 11;
+				digit1 = (digit1 == 10) ? 0 : digit1;
+				digit2 = digit2 % 11;
+				digit2 = (digit2 == 10) ? 0 : digit2;
+
+				if (Integer.parseInt(String.valueOf(cpf.charAt(9))) + Integer.parseInt(String.valueOf(cpf.charAt(10))) == (digit1 + digit2)) {
+					System.out.println("CPF válido! " + cpf);
+					response = true;
+				} else {
+					System.out.println("CPF inválido! " + cpf);
+				}
+			} else {
+				System.out.println("CPF inválido! " + cpf);
+			}
+		} catch (Exception e) {
+			System.out.println(e);
+		}
+		return response;
+	}
 }
